bundled clientside app code (for angularjs) via webpack into bundle.js too and used 'react2angular' package to run react component inside angularjs code
