+++ package.json
@@ -34,6 +34,8 @@
     "react": "^16.8.4",
     "react-big-calendar": "^0.20.3",
     "react-dom": "^16.8.4",
+    "react-router-dom": "^5.0.0",
+    "react2angular": "^4.0.6",
     "serve-favicon": "^2.5.0",
     "style-loader": "^0.23.1",
     "webpack": "^4.29.0",
