+++ public/calendar/calendar.component.js
@@ -1,5 +1,7 @@
-function calendarController(ConfigService) {}
+import angular from 'angular';
+import { react2angular } from 'react2angular';
+import CbsFullCalendar from './calendar.jsx';
 
-angular.module('calendarComponent', []).component('calendar', {
-  controller: ['ConfigService', calendarController]
-});
+angular
+	.module('calendarModule', [])
+	.component('cbsFullCalendar', react2angular(CbsFullCalendar));
