+++ public/calendar/calendar.css
@@ -1,3 +1,3 @@
-.h-90vh {
-	height: 90vh;
+.h-80vh {
+	height: 80vh;
 }
