+++ public/calendar/calendar.html
@@ -1,4 +1,1 @@
-<calendar>
-	<div id="cbs-alumni-calendar-app"></div>
-</calendar>
-<script src="dist/bundle.js"></script>
+<cbs-full-calendar></cbs-full-calendar>
