+++ public/layout/navigation.component.js
@@ -1,3 +1,5 @@
+import angular from 'angular';
+
 function navigationController(ConfigService, $http) {
 	this.selected = 'Account';
