+++ public/layout/navigation.template.html
@@ -7,7 +7,7 @@
             <li ui-sref="members" ui-sref-active="active">
                 <a href="" class="navbar-link" ng-click="$ctrl.selectNavbarItem('Directory', false);">Directory</a>
             </li>
-            <li ui-sref="calendar" ui-sref-active="active">
+            <li ui-sref="calendar({param: 'index'})" ui-sref-active="active">
                 <a href="" class="navbar-link">Calendar</a>
             </li>
         </ul>
