+++ webpack.config.js
@@ -1,13 +1,13 @@
 
 const config = {
-  entry: __dirname + '/public/calendar/react-calendar.jsx',
+  entry: __dirname + '/public/app.js',
   output: {
     path: __dirname + '/public/dist',
     filename: 'bundle.js'
   },
   module : {
     rules: [{
-      test: /\.(jsx)$/,
+      test: /\.(js|jsx)$/,
       exclude: /(node_modules|bower_components)/,
       use: {
         loader: "babel-loader"
